src/test/java/com/martzatech/vdhg/crmprojectback/infrastructure/apis/v1/controllers/RegistrationTypeControllerTests.java
package com.martzatech.vdhg.crmprojectback.infrastructure.apis.v1.controllers;

import com.martzatech.vdhg.crmprojectback.domains.customers.models.RegistrationType;
import com.martzatech.vdhg.crmprojectback.domains.customers.services.RegistrationTypeService;
import com.martzatech.vdhg.crmprojectback.infrastructure.apis.v1.mappers.RegistrationTypeApiMapper;
import com.martzatech.vdhg.crmprojectback.infrastructure.apis.v1.requests.RegistrationTypeRequest;
import com.martzatech.vdhg.crmprojectback.infrastructure.apis.v1.responses.RegistrationTypeResponse;
import java.util.List;
import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.Mockito;
import org.mockito.junit.jupiter.MockitoExtension;

@ExtendWith(MockitoExtension.class)
public class RegistrationTypeControllerTests {

  private static final Integer ANY_ID = 1;

  @InjectMocks
  private RegistrationTypeController tested;

  @Mock
  private RegistrationTypeService service;

  @Mock
  private RegistrationTypeApiMapper mapper;

  @Mock
  private RegistrationTypeRequest request;

  @Mock
  private RegistrationType model;

  @Mock
  private RegistrationTypeResponse response;

  @Test
  void whenSave_thenCallService() {
    Mockito.when(mapper.requestToModel(request)).thenReturn(model);
    Mockito.when(service.save(model)).thenReturn(model);
    Mockito.when(mapper.modelToResponse(model)).thenReturn(response);

    final RegistrationTypeResponse actual = tested.save(request);

    Assertions.assertEquals(response, actual);
  }

  @Test
  void whenFindAll_thenCallService() {
    Mockito.when(service.findAll()).thenReturn(List.of(model));
    Mockito.when(mapper.modelsToResponseList(List.of(model))).thenReturn(List.of(response));

    final List<RegistrationTypeResponse> actual = tested.findAll();

    Assertions.assertEquals(List.of(response), actual);
  }

  @Test
  void whenFindById_thenCallService() {
    Mockito.when(service.findById(ANY_ID)).thenReturn(model);
    Mockito.when(mapper.modelToResponse(model)).thenReturn(response);

    final RegistrationTypeResponse actual = tested.findById(ANY_ID);

    Assertions.assertEquals(response, actual);
  }

  @Test
  void whenDeleteById_thenCallService() {
    Assertions.assertDoesNotThrow(() -> tested.deleteById(ANY_ID));

    Mockito.verify(service).deleteById(ANY_ID);
  }
}
